+++ CHANGES.md
@@ -7,6 +7,7 @@ unreleased
     - Improve recovery of mutually recursive definitions (#1962, #1963, fixes #1953)
   + vim plugin
     - Fix error when `:MerlinOccurrencesProjectWide` fails to gather code previews (#1970)
+    - Add more short-paths tests cases (#1904)
 
 merlin 5.5
 ==========